Update README.md
Modules and their requirements :-
User Module :
1.Login
2.Registration
3.View Menu
4.Book Appointment.
Admin Module :
1.Login
2.View Menu
3.View All User
4.View All Doctors
5.View All Appointment
Doctor Module :
1.Login
2.Registration
3.View Appointment.
4.Profile .
 

Software Requirements
Technology used:
Front end – JSP
Backend – MYSQL
Software:
 1.IDE - Netbeans 8.2
 2.Database support - MySQL 5.7
 3.Operating system – Windows 8 and above
 4.Server deployment - Glassfish server
Technology:
HTML is integrated in JSP. It provides a means to structure text based information in a document. It allows users to produce web pages that include text, graphics and
hyperlinks.
 Javascript is a scripting language which supports the development of both client and server applications. It is preferred at client side to write programs that can be
executed by a web browser within the context of a web page.
 CSS(Cascading Style Sheets) is a style sheet language used for describing the presentation of a document written in a markup language. Although most often used to set the visual style of web pages and user interfaces written in HTML and XHTML, the language can be applied to any XML document,
 SQL is the language used to manipulate relational databases. It is tied closely with the relational model. It is issued for the purpose of data definition and data
manipulation.
Java Server pages is a simple yet powerful technology for creating and maintaining dynamic-content web pages. It is based on the Java programming language. It can be thought of as an extension to servlet because it provides more functionality than servlet A JSP page consists of HTML tags and JSP tags. The jsp pages are easier to maintain than servlet because we can separate designing and
development.
We require a JDBC connection between the front end and back end components to write

to the database and fetch required data